Work around race condition for endpoint connections
Since the CCAPI job to accept endpoint connections is async, we can
get into a state where the connection still shows as PENDING_ACCEPTANCE
after it's been accepted. The old wait logic looked for ACCEPTED or
PENDING, and failed out with other statuses, including
PENDING_ACCEPTANCE. This change interprets that status the same as
PENDING, and also lowers the timeout for connections (which usually
only take a minute or less) to 5 minutes.