Update terraform-provider-framework to 0.17.0
This commit updates the terraform-provider-framework and supporting
modules. Those updates include a ton of breaking changes, and thus
required a significant amount of refactoring.

The refactoring is mostly straightforward, and there should be no
functional changes. The biggest differences are around how resources
and datasources are created and configured. Still, the diff should
be easy enough to read.