[CC-8720]CMEK resource
This commit adds a new CMEK resource, which primarily manages the
configuration of customer-managed encryption keys (CMEK). CMEK is
enabled for a cluster when the resource is created, and keys can
be rotated by updating their specs.

Due to circular dependencies (new regions depend on CMEK config,
which depends on role policy, which depends on cluster), new
regions must be managed by the CMEK resource. That means the cluster
resource and CMEK resource need to keep track of which regions they
"own" and reconcile that with updates and the actual state of the
cluster. That's the bulk of the complexity here.

Destroying a CMEK resource is a no-op, since CMEK cannot be disabled.
That means that a user could be left with "floating" regions.
I need to figure out a way to block targeted destroy operations on
CMEK resources while allowing them to be removed from state when the
parent cluster is destroyed. We could attempt to remove additional
regions on destroy, but that would take significantly longer.

Happy to do a walkthrough and discuss tradeoffs at a watercooler.