Connection string data source (#73)
* Cluster cert data source

Created a new cockroach_cluster_cert data source, which downloads
and prints the TLS certificate for a given cluster.

* Connection string data source

Added a new "cockroach_connection_string" data source, which takes
in certain connection params and outputs a connection string, as
well as a list of individual connection params. Passwords are handled
locally to avoid sending sensitive data over the wire in a URL param.

Updated the Go SDK to make use of the new connection string API
endpoint.