Make SQL User password optional and generate of omitted
Previously, password was a required field for SQL users. This caused
import problems because we can't read and set the password (nor do
we want to).

This change makes it optional, and if it's omitted, we randomly
generate a 32-character random password. This password isn't saved
to state, though. The user needs to then change it via the web UI.
Setting the password field to null after the resource is created
will not change the password, but makes Terraform forget it.