Add Terraform Provider to Grant roles (#94)
The added provider takes in a user id and a set of roles
and then grants those roles to that user.

Note that for CC, persons, servicea accounts and groups
are all valid users. That is why we have to use user ids
for now until a better addressing scheme is decided on
for all user types.