Add support for resource limits
The Cockroach Cloud API now supports a way to limit the number of request
units and storage MiB that a cluster can use. This works alongside an
existing limit on the $ spend a cluster can incur. A caller can choose
to either specify resource limits or a spend limit, but not both. If
neither is specified, then the cluster will not have any limits enforced.