Send public network visibility when not private
Previously on cluster creation, if the network_visibility was not
explicitly set to private than, the unspecified or nil value was sent.
This was problematic because Azure clusters require network_visibility
of public to be passed.

Release Note: During cluster creation, explicitly send
network_visibility of public if network_visibility of false is not
set.