Use pagination with List* API endpoints for Read operations
[issue #121](https://github.com/cockroachdb/terraform-provider-cockroach/issues/121)
revealed a problem with resources where we need to use a List endpoint to
read the resource state. Namely, there is a default page size and we aren't
using pagination. That results in state getting deleted because the provider
thinks the resource has disappeared.

The customer who discovered the issue already fixed it for SQL users. This
commit fixes it for databases, IP allowlist entries, and user role grants.
It also bumps the default page size for SQL users.

Also fixes the cluster resource acceptance tests, since the supported versions
have changed.