[CC-25143] Contributor guide
Added a contributor guide section to DEVELOPMENT.md detailing how to
run and add tests, perform a release, etc. Includes TF doc links, a
general overview, and a brief Q&A with questions I've been asked
recently.

Also updated the commit template with a checklist. I often forget
to regenerate docs, so hopefully this will help. Someday, we'll have
linters for that, but this'll work for now.