[CC-25231] Generated documentation is comprehensive and accurate
Went over all the resource and database schemas and updated the
Description/MarkdownDescription fields where missing or needed
improvement. Standardized on adding a period to the end of each
description.

Added a helper to format allowed enum value lists, which show up
as bullet points in a markdown renderer.

Fixed the formatting of a few ID fields, where angle brackets were
getting swallowed by the markdown renderer.

Renamed "cockroach_cluster_data_source" to "cluster_data_source"
for consisdency and "networking_resource" to "allowlist_resource"
to match the name of the actual resource.