defer to base retry check when response is nil (#168)
We use a custom retryable http client to retry GET requests. Since it is
intended to only retry GET requests we check that the request was a GET
at the beginning of the method.  The only way to access the request at
this point in the code is through the response object, however, there are
some cases where this  object is nil.  This was resulting in a
nil pointer exception.

After this change, any errors that are seen without a set Response object
will fall through to the default handler.