support backwards-compatible serverless attribute
While a previous commit removed support for the `serverless` attribute,
this commit restores it, in order to preserve backwards-compatibility.
However, it is marked as deprecated and will be removed in a future
release of the provider.

This commit strengthens the tests to handle the many possible
transitions that are now possible.