log-export: make group specific redact noncomputed (#174)
* log-export: make group specific redact noncomputed

Currently, the "redact" field in the individual log groups is declared as
a "computed" field. When a field is marked as "computed", user removing
the field/commenting it out is sometimes not recognised as a "diff" in
terraform state. This makes it look like the "redact" field is not working

* update the CHANGELOG