rename shared attribute to again be serverless (#177)
* update provider to support latest Cloud 2.0 API

This includes shared => serverless and request_unit_rate_limit =>
provisioned_capacity renames.

* rename shared attribute to again be serverless

Previous changes added a new "shared" attribute that was intended to
replace the existing "serverless" attribute (which would have been
deprecated). The Product team decided to stick with the "serverless"
naming, so this PR switches back to using that attribute. It will no
longer be marked as deprecated (though the "spend_limit" attribute
will continue to be marked as deprecated). There should be no customer
impact, since this work is in a branch and has never been publicly
released.