[CC-27365] add user_role_grant resource
Previously, the provider supported user role management but only in an
authoritative way.  Using the user_role_grants resource would overwrite
all roles for a user.  This caused churn for terraform users trying to
manage user roles across multiple projects or via the console UI.  We
now add a resource, user_role_grant which allows management of a single
user role grant.  The existing user_role_grants resource is left to
maintain the previous functionality for those that need it.

As part of this changes, I've also renamed the role_resource files which
contain the user_role_grants resource after the full name of the role
since that seems to be the convention.

Additionally, cloud SDK is Update to 1.9.0