rebase Cloud 2.0 branch on main with latest SDK
The Cloud 2.0 branch has diverged from the main branch. This commit rebases
the Cloud 2.0 branch on main. The individual Cloud 2.0 commits no longer
build on their own, because they've been rebased on the `main` version of
the SDK. This commit pulls in the Cloud 2.0 version of the SDK to address
that issue.

Release note: None