migrate testing framework to terraform-plugin-testing
Previously, the terraform provider utilized the sdkv2 framework. In order to
stay up to date with testing framework improvements we upgrade to the new
terraform-plugin-testing framework following these instructions:

https://developer.hashicorp.com/terraform/plugin/testing/migrating

The change that instigated this upgrade is that the id field is no longer
required in the new framework.  During development of the role_grant_resource
an ID was not used because the ccapi does not provide it.  Rather than working
around this using a synthetic ID, we upgrade.

Included with this migration are various changes including:
* updating github action versions
* update code to use new libraries provided by the updated framework.
* update to use go 1.21
* fixed some broken examples
* integration test changes required by the operation of the new framework.
  Namely there are different amounts of Read calls made by the new framework.
  Some tests had call counts encoded into them and need to be updated.