make "cluster.plan" attribute writable to support plan upgrades
Make the "plan" attribute on the "cluster" resource writable so that it
can be set by users who intend to upgrade/downgrade the plan. Keep the
attribute as computed so that the system will set the value if it is not
provided by the user.

Release note: Support upgrade of plans via changes to the "plan" attribute
on the "cluster" resource.