[CC-27470] add folder data source
This change adds the folder data source and associated tests. There was
previously a folder resource only.  The new data source can find a
folder using either a path string or the ID of the folder.  For example:

    data "cockroach_folder" "team1" {
      path = "/prod/team1"
    }

    data "cockroach_folder" "prod" {
      id = var.prod_folder_id
    }

A custom validator for the folder parent_id was also added as part of
the validators to provide better validation and messaging for that
attribute.