add env var controlled output of api invocations
When adding resource tests that rely on mocked api calls, it can be
difficult to determine the count and order of these calls.  This change
adds a method `traceAPICall` which prints a line with the caller and
method name. It is enabled by setting the environment variable
`TRACE_API_CALLS=1`.

More ideally, this call could be made in the api layer itself or in a
wrapper around the client but for now, calls are made manually at each
api call site.