fix allowlist_resource churn on empty name
Previously, allowlist.name attribute was not marked as computed so
values from the server would not set the state in case it was unused.
The server side default of "" would then conflict with a nil value in
the state and show that it needed to be updated on every apply. We now
mark the allowlist name as computed so that it can update the state when
it is not set.  Additionally we stop sending the value on updates unless
it explicitly set in the resource.