private_endpoints_service: wait for all regions
Previously, when waiting for private endpoint service creation, the code
would list the known private endpoint services and ensure they were
all created.  It was unclear if this endpoint would return all the
regions at once or if it could return a subset of the regions as they
were spinning up.  We now loop over all of the cluster.Regions to ensure
every service is in the correct state before continuing.