[CC-29279] improve user_role_grant/s testing with sa user
Previously, the top level terraform user was used for testing in the
user_role_grant_resource and user_role_grants_resource tests. Changes to
this user are fragile and could affect other tests or require manual
fixing on failure. This commit updates the user associated with those
tests to be a newly created service account user. Additionally this
commit also includes a new acceptance test for
user_role_grants_resource.