[CC-29279] fix failing cluster creation test for Terraform v1.9.4
The terraform testing framework uses the terraform binary to run the
tests. If the path of an existing binary is not passed in, it downloads
the most recent version. TestIntegrationDedicatedClusterResource has
been failing in CI due to updates introduced in the most recent versions
of terraform. Upon research, the expectation now is that the order and
number of API calls might change with any/each terraform update. The fix
in this commit gets the test working with Terraform v1.9.4.