[CC-29279] improve user_role_grant/s testing with sa user (#225)
* [CC-29279] improve user_role_grant/s testing with sa user

Previously, the top level terraform user was used for testing in the
user_role_grant_resource and user_role_grants_resource tests. Changes to
this user are fragile and could affect other tests or require manual
fixing on failure. This commit updates the user associated with those
tests to be a newly created service account user. Additionally this
commit also includes a new acceptance test for
user_role_grants_resource.

* [CC-29279] fix failing cluster creation test for Terraform v1.9.4

The terraform testing framework uses the terraform binary to run the
tests. If the path of an existing binary is not passed in, it downloads
the most recent version. TestIntegrationDedicatedClusterResource has
been failing in CI due to updates introduced in the most recent versions
of terraform. Upon research, the expectation now is that the order and
number of API calls might change with any/each terraform update. The fix
in this commit gets the test working with Terraform v1.9.4.