pass Plan to CreateCluster if it was specified by the user
Previously, the Plan field was not passed to the CreateCluster API, even if
it was explicitly set by the user. This prevented the server from validating
the cluster configuration. For example, the server was not able to return
an error when Plan=BASIC and provisioned_vcpus are set. The fix is to pass
the Plan field to CreateCluster if the user explicitly sets it.