Update to version 2.0.1 of cockroach-cloud-sdk-go
This was a breaking change update.

The main required fix was in metric_export_prometheus_config_resource.go
to reduce the number of parameters being passed into
EnablePrometheusMetricExport.

Additionally, since the ApiOidcConfig endpoints are now deprecated, I
had to explicitly bypass the lint rule for those lines while they are
still in use.  The deprecation/removal of that resource will come in a
later PR.