remove waitForClusterLock in favor of waitForClusterReady
Previously waitForClusterLock would check in the state to see if the
cluster was locked and if so it would call waitForClusterReady.  This
relied on the cluster state being up to date with the server. Normally,
if all operations are occuring within terraform this could usually be
relied on but due to a bug in the ccapi, it looks like waiting for
successful completion of shutting down the prometheus metric export is
broken.

While a separate fix for the metric export issue will eventually be
made, we go ahead and remove reliance on the current tf state to protect
against issues like this in the future.

Note that waitForClusterReady waits for any ready state where as
waitForClusterLock would check initially for locked state only.  In
practice, wherever waitForClusterLock would be used, any ready state
is generally acceptable.  Also, note that the implementation of
waitForClusterLock was falling back to waitForClusterReady anyways.

A note on alternative considerations: A more robust approach here would
be to try the operations directly and retry on certain failures such as
locked.  The removes the critical period in between when we've found
cluster to be unlocked and the actual operation.  The current approach
was taken due to its simplicity.  Perhaps the more robust approach can
be explored at another time.