specify RetentionDays in final backupconfig test case
Previously a cluster test was creating a backup configuration that
didn't specify the RetentionDays directly.  This meant that it would
rely on the server specified default.  In the test case, it was then
compared to a local object which had a value different from the default
which was causing the test to fail.

Rather than change the local value to match the server side default, we
instead pass the local value into the test so that it is created with
the non-default value.