use 4VCPU in testDedicatedClusterResource
Previously, the testDedicatedClusterResource test was requesting 2VCPU
clusters which was failing during an actual acceptance test.  This
updates it to use 4VCPU.  We additionally had to increase the scale step
to 8VCPU to provide a real value that would necessitate an update.

additionaly, we update the default from 2 to 4 in the advanced example.