[CC-30640] allow manual version control for standard clusters
STANDARD clusters now accept the cockroach_version attribute on cluster
update for manual upgrades control.  The serverless.upgrade_type
attribute must be set to MANUAL.

As part of this an API call to validate cockraoch_version for ADVANCED
clusters prior to the update was removed.  This logic provided some
nicer error messages for advanced clusters but the logic was complex, it
required an additional api call to fetch the valid versions and standard
tier version checking is not supported.  We now rely on the serverside
validation for checking versions.