[CC-30766] disallow 0 as a value for disk_iops
0 is a valid value to pass into the api for disk_iops.  That, along with
omitting the value in the api call results in the default iops value.
However, the api returns the actual value for the iops that was chosen.
This results in a configuration conflict with the set value of 0 and the
actual value that was returned from the server.  For this reason we now
disallow 0 from being set and we clarify in the docs that omitting the
value is the correct way to ask for the default.