[CC-30768] don't allow node_count for serverless cluster
Previously it was possible to pass in the node_count value for
serverless region because the type is shared with dedicated clusters.
This ended up causing a state inconsistency however becase node_count is
coerced to 0 by the api which conflicts with a value added in the
config. To solve this, we simply validate that node_count is not used
unless cluster plan is Advanced.