CC-30971: update state with allowlist api responses
Previously the object the api returned for created and updated allowlist
entries was ignored when updating the terraform state.  This meant that
defaults set on the server side and returned were not correctly being
set and could cause state inconsistencies in case they were different.
Specifically there was a case where not setting an initial value for the
create would return an empty string.  The empty string caused a state
consistency error with the unset value.
To fix this we now set the state based on the response.  As part of this
the code was reworked to be more similar to how other resources are
structured.

Tests were added to show that creating an allowlist without a name now
succeeds.

Also, as part of this change, instances of allowList were renamed to
allowlist for consistency.