Install terraform for github actions that need it
The image in use for our ci actions, ubuntu-latest is now defaulting to
ubuntu-24.04.  In the previous version, terraform was installed by
default but has now been removed so we install it manually for each job
that relies on it.

https://github.com/actions/runner-images/issues/10636

We choose not to pin the version of terraform that is installed so we
are always testing against the the latest CLI version. If there is a
discrepancy between versions that only affects tests, developers can
update to the latest version and ensure the tests work for that.