add name back to allowlist creation
In a previous commit (272cdd80b3f93b61facaffa39a886dcd368980d9), passing
the allowlist name during creation of that resource was mistakenly
removed.  The test exercising it was also changed such that it didn't
catch that issue.  We now add back the name and augment the test so that
it covers this case again.

closes #263