[FEB-93] add cluster attribute for full version string
In order to orchestrate version upgrades we allow specifying the major
version of the cluster in the cluster resource.  Originally the cluster
data source exposed the full version string but having this value be
a different format from the resource (major vs full) was causing
confusion so the data source was changed to match the resource.

At this point, the full version string was not available in any resource
or data source so we add it back here as "full_version".