handle external deletion of service account in user role grant resource
Previously, the external deletion of a service account resource managed
in Terraform caused errors on apply if there was a user role grant
resource that referenced the user ID associated with the deleted service
account. This PR fixes this issue by adding a RequireReplace plan
modifier on the UserID attribute of the UserRoleGrant resource.

When the service account was deleted externally, Terraform automatically
created a new service account on apply. However, the existing user role
grant resource in Terraform was still referencing the user ID of the
deleted service account in the state file, leading to a mismatch between
the state and the data returned by the API. This mismatch caused the
apply to fail.

Closes: CLOUDOPS-13734