Add egress_private_endpoint resource
This commit implements an egress private endpoint as a Terraform
resource. See below for official Cockroach Cloud documentation for the
feature. This feature is currently in limited access and is not yet
available to all customers.

Note that support for adding custom domain names to an egress private
endpoint will come in a later commit.

https://www.cockroachlabs.com/docs/cockroachcloud/egress-private-endpoints.html