Add egress_private_endpoint_domain_names resource
This commit implements custom domain names for an egress private
endpoint as a Terraform resource. See below for official Cockroach
Cloud documentation for the feature. This feature is currently in
limited access and is not yet available to all customers.

This resource is separate from the existing egress private endpoint
resource for sequencing reasons: in order for domain names to be
successfully added to an egress private endpoint, the endpoint must be
in the AVAILABLE state. However, it is possible that other resources
might be needed to make the endpoint AVAILABLE. Therefore, we have two
resources, to allow Terraform to create one separately from the other.

https://www.cockroachlabs.com/docs/cockroachcloud/egress-private-endpoints.html