[CNSL-1406] Expose private_endpoint_dns in cluster region list
Previously, `private_endpoint_dns` was not exposed in the cluster region
list. Since it is used to connect a cluster to GCP Private Service
Connect, customers needed to manually retrieve the `private_endpoint_dns`
from the CockroachCloud UI or API, breaking end-to-end automation.

This change adds `private_endpoint_dns` to the cluster region list to
allow customers to directly access it via Terraform.

This commit builds on #281.

Co-authored-by: Miles Sorlie <msorlie@protonmail.com>